refactor so mapcreator  is in a separate mapController
